[backend] flexible backend
i've added extra configuration for the backend - you can now specify
which modules you WANT to be run and where all modules are located.

but this is a kind of breaking change, since the `modules` folder now
needs a new file in it - `modules.json`.

its structure goes as follows: ```
[
  {
    "name": "Module Name",
    "colour": "<ANSI_ESCAPE_CODE>",
    "file": "<FILE_LOCATION"
  }
]```

it should look like `[Module Name] blah blah blah`!

this feature allows to get rid of `fs` (which i will do in the next
commit), and i think it also saves on I/O